Add files via upload
Made using Unity3D. The ball moves in the direction you point at. Media Pipe library is used in python for getting coordinates of the finger. UDP Server is made inside unity and client inside python to communicate the Direction info. The coordinatess of the pointer finger are used to obtain the direction vector which is then used to direct the ball. Particles have been used . Navigation AI is used in enemy.